pyproject.toml
[project]
name = "mini-mancer"
version = "1.0.0a1"
description = "Sophisticated AI Bot Creation System with OpenServ workflow integration"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agno>=0.5.50",
    "fastapi>=0.115.13",
    "httpx>=0.28.1",
    "openai>=1.0.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.0",
    "python-telegram-bot>=22.1",
    "uvicorn>=0.34.3",
    "sqlalchemy>=2.0.0",
    "psycopg[binary]>=3.0.0",
    "psutil>=7.0.0",
    "types-psutil>=7.0.0.20250601",
    "telethon>=1.40.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "pytest>=8.4.1",
]
test = [
    "psutil>=7.0.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
]
quality = [
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "bandit>=1.7.0",
    "pre-commit>=3.5.0",
    "isort>=5.12.0",
    "black>=23.0.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings  
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "PL",   # pylint
]
ignore = [
    "PLR0913",  # Too many arguments to function call
    "PLR0915",  # Too many statements
    "PLR2004",  # Magic value used in comparison
    "E501",     # line too long (handled by black)
    "E203",     # whitespace before ':' (conflicts with black)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

[tool.ruff.lint.isort]
known-first-party = ["src"]
force-single-line = false
lines-after-imports = 2

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "agno.*",
    "telegram.*",
    "psycopg.*",
]
ignore_missing_imports = true

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
src_paths = ["src", "tests"]
known_first_party = ["src"]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.uv.workspace]
members = ["bootstrap"]
